Move reduction operations out of dataset.h
