+++ dataset/CMakeLists.txt
@@ -3,8 +3,8 @@
 set(TARGET_NAME "scipp-dataset")
 set(INC_FILES
     include/scipp/dataset/dataset.h include/scipp/dataset/dataset_index.h
-    include/scipp/dataset/except.h include/scipp/dataset/counts.h
-    include/scipp/dataset/sort.h
+    include/scipp/dataset/reduction.h include/scipp/dataset/except.h
+    include/scipp/dataset/counts.h include/scipp/dataset/sort.h
 )
 
 set(SRC_FILES
@@ -19,6 +19,7 @@ set(SRC_FILES
     histogram.cpp
     map_view.cpp
     operations.cpp
+    reduction.cpp
     sort.cpp
     string.cpp
     unaligned.cpp
