+++ dataset/include/scipp/dataset/dataset.h
@@ -867,16 +867,6 @@ SCIPP_DATASET_EXPORT DataArray astype(const DataArrayConstView &var,
 SCIPP_DATASET_EXPORT Dataset merge(const DatasetConstView &a,
                                    const DatasetConstView &b);
 
-SCIPP_DATASET_EXPORT DataArray flatten(const DataArrayConstView &a,
-                                       const Dim dim);
-SCIPP_DATASET_EXPORT Dataset flatten(const DatasetConstView &d, const Dim dim);
-
-SCIPP_DATASET_EXPORT DataArray sum(const DataArrayConstView &a, const Dim dim);
-SCIPP_DATASET_EXPORT Dataset sum(const DatasetConstView &d, const Dim dim);
-
-SCIPP_DATASET_EXPORT DataArray mean(const DataArrayConstView &a, const Dim dim);
-SCIPP_DATASET_EXPORT Dataset mean(const DatasetConstView &d, const Dim dim);
-
 SCIPP_DATASET_EXPORT DataArray concatenate(const DataArrayConstView &a,
                                            const DataArrayConstView &b,
                                            const Dim dim);
