+++ python/operations.cpp
@@ -15,24 +15,6 @@ using namespace scipp::dataset;
 
 namespace py = pybind11;
 
-template <class T> void bind_flatten(py::module &m) {
-  m.def("flatten",
-        py::overload_cast<const typename T::const_view_type &, const Dim>(
-            &flatten),
-        py::arg("x"), py::arg("dim"), py::call_guard<py::gil_scoped_release>(),
-        Docstring()
-            .description("Flatten the specified dimension into event lists, "
-                         "equivalent to summing dense data.")
-            .raises("If the dimension does not exist, or if x does not contain "
-                    "event lists.")
-            .seealso(":py:func:`scipp.sum`")
-            .returns("The flattened data.")
-            .rtype<T>()
-            .template param<T>("x", "Data container to flatten.")
-            .param("dim", "Dimension over which to flatten.", "Dim")
-            .c_str());
-}
-
 template <class T> void bind_concatenate(py::module &m) {
   auto doc = Docstring()
                  .description(R"(
@@ -136,10 +118,6 @@ template <typename T> void bind_contains_events(py::module &m) {
 }
 
 void init_operations(py::module &m) {
-  bind_flatten<Variable>(m);
-  bind_flatten<DataArray>(m);
-  bind_flatten<Dataset>(m);
-
   bind_concatenate<Variable>(m);
   bind_concatenate<DataArray>(m);
   bind_concatenate<Dataset>(m);
